+++ packages/editor/src/components/collab-sidebar/add-comment.js
@@ -35,7 +35,7 @@ export function AddComment( {
 			const selectedBlock = getSelectedBlock();
 			return {
 				clientId: selectedBlock?.clientId,
-				blockCommentId: selectedBlock?.attributes?.metadata?.commentId,
+				blockCommentId: selectedBlock?.attributes?.metadata?.noteId,
 				isEmptyDefaultBlock: selectedBlock
 					? isUnmodifiedDefaultBlock( selectedBlock )
 					: false,